feat: move /ratelimits route into the / route
