+++ oprish/src/routes/mod.rs
@@ -1,17 +1,21 @@
 pub mod messages;
-pub mod rate_limits;
 
 use rocket::{serde::json::Json, Route, State};
 use rocket_db_pools::Connection;
-use todel::{http::ClientIP, models::InstanceInfo, Conf};
+use todel::{
+    http::ClientIP,
+    models::{InstanceInfo, InstanceRateLimits},
+    Conf,
+};
 
 use crate::{
     rate_limit::{RateLimitedRouteResponse, RateLimiter},
     Cache, VERSION,
 }; // poggers
 
-#[get("/")]
+#[get("/?<rate_limits>")]
 pub async fn index(
+    rate_limits: bool,
     address: ClientIP,
     mut cache: Connection<Cache>,
     conf: &State<Conf>,
@@ -28,9 +32,14 @@ pub async fn index(
         effis_url: &conf.effis.url,
         file_size: conf.effis.file_size,
         attachment_file_size: conf.effis.attachment_file_size,
+        rate_limits: rate_limits.then_some(InstanceRateLimits {
+            oprish: conf.oprish.rate_limits.clone(),
+            pandemonium: conf.pandemonium.rate_limit.clone(),
+            effis: conf.effis.rate_limits.clone(),
+        }),
     }))
 }
 
 pub fn get_routes() -> Vec<Route> {
-    routes![index, rate_limits::rate_limits]
+    routes![index]
 }
