+++ todel/src/models/info.rs
@@ -1,3 +1,4 @@
+use crate::conf::{EffisRateLimits, OprishRateLimits, RateLimitConf};
 use serde::{Deserialize, Serialize};
 
 /// The instance info payload
@@ -13,6 +14,8 @@ pub struct InstanceInfo {
     pub effis_url: String,
     pub file_size: u64,
     pub attachment_file_size: u64,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub rate_limits: Option<InstanceRateLimits>,
 }
 
 /// The instance info payload
@@ -28,4 +31,14 @@ pub struct InstanceInfo<'a> {
     pub effis_url: &'a str,
     pub file_size: u64,
     pub attachment_file_size: u64,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub rate_limits: Option<InstanceRateLimits>,
+}
+
+/// The type which represents all of an instance's rate limit configuration
+#[derive(Debug, Clone, Serialize, Deserialize)]
+pub struct InstanceRateLimits {
+    pub oprish: OprishRateLimits,
+    pub pandemonium: RateLimitConf,
+    pub effis: EffisRateLimits,
 }
